Refactor: rename main components
