+++ src/components/Main/ExchangeRate/ExchangeRateChart/style.js
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 
-export const ExchRateChartWrapper = styled.div`
+export const ExchangeRateChartWrapper = styled.div`
   /* border: 1px solid black; */
   width: 100%;
   height: 100%;
