+++ src/components/Main/ExchangeRate/style.js
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 
-export const ExchRate = styled.div`
+export const ExchangeRateWrapper = styled.div`
   width: 100%;
   height: 100%;
   background: #fff;
