+++ src/components/Main/MajorStock/style.js
