+++ src/components/Main/StockIndex/index.jsx
@@ -1,8 +1,11 @@
 import React, { useCallback, useEffect, useRef, useState } from "react";
+import { StockIndexWrapper } from "./style";
 import * as d3 from "d3";
-import { ExchRateChartWrapper } from "./style";
 import useResizeObserver from "@utils/useResizeObserver";
 import useDebounce from "@utils/useDebounce";
+import stock_up from "@images/stock_up.svg";
+import stock_down from "@images/stock_down.svg";
+import stock_none from "@images/stock_none.svg";
 
 const data = [
   { stock: 2900, date: "10:00" },
@@ -43,10 +46,10 @@ const data = [
   { stock: 2900, date: "12:50" },
 ];
 
-const ExchRateChart = () => {
-  const exchRateChartRef = useRef(null);
+const StockIndex = ({ name }) => {
   const svgRef = useRef(null);
-  const size = useResizeObserver(exchRateChartRef);
+  const stockIndexRef = useRef(null);
+  const size = useResizeObserver(stockIndexRef);
   const resize = useDebounce(size, 200);
 
   //tick 분할 함수
@@ -80,19 +83,19 @@ const ExchRateChart = () => {
     }
 
     //초기 셋팅
-    const margin = { top: 30, right: 50, bottom: 30, left: 0 };
+    const margin = { top: 50, right: 50, bottom: 50, left: 50 };
     const width = resize.width - (margin.left + margin.right);
-    const height = 150;
-    const xTickCount = 6;
-    const yTickCount = 5;
-    const xTickBlankCount = 0;
+    const height = 100;
+    const xTickCount = 4;
+    const yTickCount = 4;
+    const xTickBlankCount = 2;
+    const eveStock = 2700;
     const minStock = d3.min(data.map((d) => d.stock)) - 300;
     const maxStock = d3.max(data.map((d) => d.stock)) + 300;
-    const xLabel = "(회차)";
-    const yLabel = "(원)";
 
     //svg 셋팅
     const svg = d3.select(svgRef.current);
+
     svg
       .attr("width", width + margin.left + margin.right)
       .attr("height", height + margin.top + margin.bottom);
@@ -105,9 +108,7 @@ const ExchRateChart = () => {
     const xAxis = d3
       .axisBottom(xScale)
       .tickValues(setTickCount(0, data.length, xTickCount, "center"))
-      .tickFormat((index) => data[index]["date"])
-      .tickSize(10);
-
+      .tickFormat((index) => data[index]["date"]);
     svg
       .selectAll(".x-axis")
       .style(
@@ -115,17 +116,7 @@ const ExchRateChart = () => {
         `translate(${margin.left}px, ${height + margin.top}px)`,
       )
       .style("stroke-opacity", 0)
-      .call(xAxis)
-      .call((g) =>
-        g
-          .selectAll(".x-label")
-          .append("text")
-          .attr("class", "x-label")
-          .attr("text-anchor", "start")
-          .attr("fill", "black")
-          .attr("transform", `translate(${width + 13}, 20)`)
-          .text(xLabel),
-      );
+      .call(xAxis);
 
     //x축 line
     const xScaleLine = d3
@@ -135,7 +126,6 @@ const ExchRateChart = () => {
     const xAxisLine = d3
       .axisBottom(xScaleLine)
       .tickValues(setTickCount(0, data.length, xTickCount * 2, "center"))
-      .tickSize(0)
       .tickFormat("");
     svg
       .selectAll(".x-axis-line")
@@ -143,7 +133,7 @@ const ExchRateChart = () => {
         "transform",
         `translate(${margin.left}px, ${height + margin.top}px)`,
       )
-      .style("stroke-opacity", 1)
+      .style("stroke-opacity", 0)
       .call(xAxisLine)
       .call((g) =>
         g
@@ -153,41 +143,57 @@ const ExchRateChart = () => {
           .style("stroke-opacity", 1),
       );
 
-    //y축 right + line
+    //y축 left + line
     const yScale = d3
       .scaleLinear()
       .domain([minStock, maxStock])
       .range([height, 0]);
     const yAxis = d3
-      .axisRight(yScale)
-      .tickValues(setTickCount(minStock, maxStock, yTickCount))
-      .tickSize(10);
+      .axisLeft(yScale)
+      .tickValues(setTickCount(minStock, maxStock, yTickCount));
     svg
       .selectAll(".y-axis")
       .call(yAxis)
-      .style(
-        "transform",
-        `translate(${width + margin.left}px, ${margin.top}px)`,
-      )
+      .style("transform", `translate(${margin.left}px, ${margin.top}px)`)
       .style("stroke-opacity", 0)
       .call((g) =>
         g
           .selectAll(".tick line")
           .style("transform", "translateX(0px)")
-          .attr("x2", -width)
+          .attr("x2", width)
           .style("stroke-opacity", 1),
-      )
-      .call((g) =>
-        g
-          // .selectAll(".x-label > *")
-          .selectAll(".y-label")
-          .append("text")
-          .attr("class", "y-label")
-          .attr("text-anchor", "start")
-          .attr("fill", "black")
-          .attr("transform", `translate(13, -15)`)
-          .text(yLabel),
       );
+
+    //y축 right
+    const yAxisRight = d3
+      .axisRight(yScale)
+      .tickValues(setTickCount(minStock, maxStock, yTickCount));
+    svg
+      .selectAll(".y-axis-right")
+      .call(yAxisRight)
+      .style(
+        "transform",
+        `translate(${width + margin.left}px, ${margin.top}px)`,
+      )
+      .style("stroke-opacity", 0);
+
+    //x축 today line
+    const xScaleToday = d3
+      .scaleLinear()
+      .domain([0, data.length - 1])
+      .range([0, width]);
+    const todayLine = d3
+      .line()
+      .x((value, index) => xScaleToday(index))
+      .y((value) => yScale(value));
+    svg
+      .selectAll(".todayLine")
+      .data([data.map((data) => eveStock)])
+      .join("path")
+      .style("transform", `translate(${margin.left}px, ${margin.top}px)`)
+      .attr("class", "todayLine")
+      .attr("d", todayLine)
+      .style("stroke-opacity", 1);
 
     //data line
     const dataLine = d3
@@ -203,26 +209,42 @@ const ExchRateChart = () => {
       .attr("class", "dataLine")
       .attr("d", dataLine)
       .attr("fill", "none");
-
-    return () => {
-      d3.selectAll(".x-label > *").remove();
-      d3.selectAll(".y-label > *").remove();
-    };
+    svg
+      .selectAll(".dataLineBlur")
+      .data([data.map((data) => data["stock"])])
+      .join("path")
+      .style("transform", `translate(${margin.left}px, ${margin.top}px)`)
+      .attr("class", "dataLineBlur")
+      .attr("d", dataLine)
+      .attr("fill", "none");
   }, [data, resize]);
 
   return (
-    <ExchRateChartWrapper ref={exchRateChartRef}>
-      <svg ref={svgRef}>
-        <g className="x-axis">
-          <g className="x-label" />
-        </g>
-        <g className="y-axis">
-          <g className="y-label" />
-        </g>
-        <g className="x-axis-line" />
-      </svg>
-    </ExchRateChartWrapper>
+    <StockIndexWrapper>
+      <div className="topInfo">
+        <h2>{name}</h2>
+        <div className="info">
+          <div className="index">
+            2,718.74
+            <span className="vs">
+              <img src={stock_up} alt="stock up" />
+              11.95
+            </span>
+            <span className="rate">+0.44%</span>
+          </div>
+          <div className="date">2022.02.23 14:15 장중</div>
+        </div>
+      </div>
+      <div className="stockIndexRef" ref={stockIndexRef}>
+        <svg ref={svgRef}>
+          <g className="x-axis" />
+          <g className="y-axis" />
+          <g className="y-axis-right" />
+          <g className="x-axis-line" />
+        </svg>
+      </div>
+    </StockIndexWrapper>
   );
 };
 
-export default ExchRateChart;
+export default StockIndex;
