+++ src/components/Main/TopStock/style.js
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
 import { hierarchy } from "d3";
 
-export const TopStock = styled.div`
+export const TopStockWrapper = styled.div`
   flex: 2 1 0;
   height: 544px;
   background: #fff;
