+++ src/routes/Domestic/Domestic.jsx
@@ -1,38 +1,38 @@
 import React, { useState, useEffect, useCallback } from "react";
 import { DomesticWrapper } from "./style";
-import MainStockIndex from "@components/Main/MainStockIndex";
-import MainHeadeer from "@components/Main/MainHeadeer";
-import MainTopStock from "@components/Main/MainTopStock/index";
-import MainExchRate from "@components/Main/MainExchRate";
-import MainMarketIndi from "@components/Main/MainMarketIndi";
+import StockIndex from "@components/Main/StockIndex";
+import RecentStock from "@components/Main/RecentStock";
+import TopStock from "@components/Main/TopStock/index";
+import ExchangeRate from "@components/Main/ExchangeRate";
+import MarketIndi from "@components/Main/MarketIndi";
 import StorageInput from "@components/Main/StorageInput";
 
 const Domestic = () => {
   return (
     <DomesticWrapper>
       <div className="row">
-        <MainHeadeer />
+        <RecentStock />
       </div>
       <div className="row">
         <div className="col">
-          <MainStockIndex name="코스피 지수" />
+          <StockIndex name="코스피 지수" />
         </div>
         <div className="col">
-          <MainStockIndex name="코스닥 지수" />
+          <StockIndex name="코스닥 지수" />
         </div>
         <div className="col">
-          <MainStockIndex name="코스피 200 지수" />
+          <StockIndex name="코스피 200 지수" />
         </div>
       </div>
       <div className="row">
         <div className="col col2">
-          <MainTopStock />
+          <TopStock />
         </div>
         <div className="col">
-          <MainExchRate />
+          <ExchangeRate />
         </div>
         <div className="col">
-          <MainMarketIndi />
+          <MarketIndi />
         </div>
       </div>
       <StorageInput type="recent" />
